requirements.txt
numpy==1.19.0
scipy==1.5.1
docopt==0.6.2
glob2==0.7
hydra-core==0.11.3
ipython==7.16.1
json5==0.9.5
jupyterlab==2.1.5
matplotlib==3.2.2
more-itertools==8.4.0
notebook==6.0.3
opencv-python==4.3.0.36
pandas==1.0.5
Pillow==7.2.0
pytest==5.4.3
scikit-image==0.17.2
scikit-learn==0.23.1
torch==1.5.1
torchvision==0.6.1
tqdm==4.47.0
